Hoist ConnectionController
